+++ src/calendar.js
@@ -8844,7 +8844,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
         if ( type === "csv" ) {
             contents = getCsvContents( contentsEvents );
         } else if ( type === "xml" ) {
-            contents = getXmlContents( contentsEvents );
+            contents = getXmlContents( contentsEvents, filename );
         } else if ( type === "json" ) {
             contents = getJsonContents( contentsEvents );
         } else if ( type === "text" ) {
@@ -9249,11 +9249,19 @@ function calendarJs( elementOrId, options, searchOptions ) {
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
      */
 
-    function getXmlContents( orderedEvents ) {
+    function getXmlContents( orderedEvents, filename ) {
         var contents = [],
             orderedEventLength = orderedEvents.length;
 
         contents.push( "<?xml version=\"1.0\" ?>" );
+        contents.push( "<Calendar>" );
+
+        if ( isDefined( filename ) ) {
+            contents.push( "<Filename>" + filename + "</Filename>" );
+        }
+
+        contents.push( "<LastModified>" + getExportDateTime() + "</LastModified>" );
+        contents.push( "</Calendar>" );
         contents.push( "<Events>" );
 
         for ( var orderedEventIndex = 0; orderedEventIndex < orderedEventLength; orderedEventIndex++ ) {